Add dealing damage overtime logic for harpoon tower
- Implement ActivateAttack() to apply a slowing effect and initial damage, then trigger periodic damage ticks over time.
- Use coroutine DamageOvertimeCo() to apply damage at regular intervals during the effect duration.
- Prevent the harpoon tower from losing its target until the full attack sequence completes.
- Add logic to reset attack state and projectile after the damage-over-time finishes.

+++ Assets/Scripts/Enemy/EnemyHeavy.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class EnemyHeavy : Enemy
 {
     [Header("Enemy Details")]
-    [SerializeField] private int shieldAmount = 50;
+    [SerializeField] private float shieldAmount = 50;
     [SerializeField] private EnemyShield shieldObject;
 
     protected override void Start()
@@ -18,7 +18,7 @@ public class EnemyHeavy : Enemy
             shieldObject.gameObject.SetActive(true);
     }
 
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(float damage)
     {
         if (shieldAmount <= 0)
             base.TakeDamage(damage);
